src/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
redis>=4.6.0,<5.0.0
fastapi-cache2[redis]>=0.2.1
pydantic>=2.5.3
